+++ src/query-tools.js
@@ -1,4 +1,5 @@
 /* eslint-disable unicorn/no-array-callback-reference, unicorn/no-fn-reference-in-iterator, unicorn/prefer-ternary */
+/** @type {Function} */
 let debug = () => {}; try { debug = require('debug')('Uttori.StorageProvider.JSON.QueryTools'); } catch {}
 const R = require('ramda');
 const { parseQueryToRamda, validateQuery, fyShuffle } = require('uttori-utilities');
@@ -8,10 +9,12 @@ const { parseQueryToRamda, validateQuery, fyShuffle } = require('uttori-utilitie
  *
  * @param {string} query - The SQL-like query to parse.
  * @param {object[]} objects - An array of object to search within.
- * @returns {object[]} Returns an array of all matched documents.
+ * @returns {object[]|number} Returns an array of all matched documents, or a count.
  * @example
+ * ```js
  * processQuery('SELECT name FROM table WHERE age > 1 ORDER BY RANDOM LIMIT 3', [{ ... }, ...]);
  * ➜ [{ ... }, ...]
+ * ```
  */
 const processQuery = (query, objects) => {
   debug('Processing Query:', query);
@@ -41,28 +44,6 @@ const processQuery = (query, objects) => {
       }),
     )(filtered);
   }
-
-  // const sorter = R.cond([
-  //   [
-  //     R.pathEq(['sort'], 'ASC'),
-  //     value =>
-  //       R.ascend(
-  //         R.prop(R.path(['prop'], value))
-  //       )
-  //   ],
-  //   [
-  //     R.pathEq(['sort'], 'DESC'),
-  //     value =>
-  //       R.descend(
-  //         R.prop(R.path(['prop'], value))
-  //       )
-  //   ],
-  //   [
-  //     R.pathEq(['sort'], 'RANDOM'),
-  //     value => R.comparator(() => Math.random() - Math.random())
-  //   ],
-  // ])({ prop: 'age', sort: 'RANDOM' });
-  // R.sortWith([sorter])(docs);
 
   // Limit
   if (limit > 0) {