+++ types/index.d.ts
@@ -1,5 +1,5 @@
 declare module "query-tools" {
-    export function processQuery(query: string, objects: object[]): object[];
+    export function processQuery(query: string, objects: object[]): object[] | number;
 }
 declare module "storage-provider" {
     export = StorageProvider;
@@ -7,11 +7,12 @@ declare module "storage-provider" {
         constructor(config: {
             content_directory: string;
             history_directory: string;
-            extension: string;
-            update_timestamps: boolean;
-            use_history: boolean;
-            spaces_document: number;
-            spaces_history: number;
+            extension?: string;
+            update_timestamps?: boolean;
+            use_history?: boolean;
+            use_cache?: boolean;
+            spaces_document?: number;
+            spaces_history?: number;
         });
         config: {
             content_directory: string;
@@ -19,18 +20,20 @@ declare module "storage-provider" {
             extension: string;
             update_timestamps: boolean;
             use_history: boolean;
+            use_cache: boolean;
             spaces_document: number;
             spaces_history: number;
         };
-        documents: any[];
-        all(): Promise<any>;
-        getQuery(query: string): Promise<any>;
-        get(slug: string): Promise<any>;
-        getHistory(slug: string): Promise<any>;
+        refresh: boolean;
+        documents: {};
+        all(): object;
+        getQuery(query: string): Promise<UttoriDocument[] | number>;
+        get(slug: string): Promise<UttoriDocument | undefined>;
+        getHistory(slug: string): Promise<string[]>;
         getRevision({ slug, revision }: {
             slug: string;
-            revision: number;
-        }): Promise<any>;
+            revision: string | number;
+        }): Promise<UttoriDocument | undefined>;
         add(document: UttoriDocument): Promise<void>;
         private updateValid;
         update({ document, originalSlug }: {
@@ -39,7 +42,6 @@ declare module "storage-provider" {
         }): Promise<void>;
         delete(slug: string): Promise<void>;
         updateHistory(slug: string, content: string, originalSlug?: string): Promise<void>;
-        refresh(): Promise<void>;
     }
     namespace StorageProvider {
         export { UttoriDocument };